git-annex-shell: accept repo paths without '~/'
Using git-annex 5.20140717, Gitolite was throwing the followin error
when running 'git annex whereis file.txt' with a Gitolite remote:

  FATAL: bad git-annex-shell command: git-annex-shell 'configlist' \
  '/repo.git' at /usr/bin/commands/git-annex-shell line 25, <DATA> \
  line 1.

This is due to the regex Gitolite uses to validate git-annex-shell
commands. It only accepts repository paths starting with '/~/', but it
seems that (some versions of) git-annex use repository paths that simply
start with '/'.